Rephrase how we determine our IP address.
The bootloader wants to use multicast to communicate with the PCM. So
we need to know the IP address of our endpoint. It is not obbvious
that there is a proper POSIX method to do this. I had used a Linux-y
/proc solution, but switched to the python netifaces module. It is
another external requirement, but has to be better than homebrewing.