Add arm-> motor indirection.
Basically, the NIR FPA arms and motor ids do not correspond simply. We
have quite a bit of existing metrology, etc. so want to keep the
existing user-level IDs. where arms "a,b,c" and "ccdMotor[123]" match
up and are not changed.

The vertical arms are A, and the other two are clockwise from that,
facing the window form the outside.

But now, internally and just for the controller commands and
responses, we map to the motor controller axis number from the arm
number.

Try to use "arm" and "motor" instead of "axis".